% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bochner.R
\name{bochner}
\alias{bochner}
\title{Compute Bochner's Theorem}
\usage{
bochner(acf, delta = 1, bias = FALSE)
}
\arguments{
\item{acf}{acf of the time-series}

\item{delta}{sampling interval}

\item{bias}{boolean value to bias as per Sykulski (2019)}
}
\value{
A vector of the positive spectral density,
excluding zero and Nyquist frequencies
}
\description{
Compute Bochner's Theorem
}
