% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwelch.R
\name{dwelch}
\alias{dwelch}
\title{Calculate the Debiased Welch Estimator (slow)}
\usage{
dwelch(
  ts,
  m,
  l,
  s,
  k = NULL,
  delta = 1,
  h = NULL,
  model = c("vanilla", "nnls"),
  centres = NULL,
  widths = NULL,
  lowers = NULL,
  uppers = NULL
)
}
\arguments{
\item{ts}{time-series as a vector}

\item{m}{number of blocks}

\item{l}{length of blocks}

\item{s}{shift factor, alternatively specified by the overlap parameter}

\item{k}{number of bases}

\item{delta}{sampling interval}

\item{h}{optional window to apply to data. Defaults to rectangular}

\item{model}{what solution algorithm to use. "vanilla" computes
the standard WLS esitmate with. "nnls" constrains the solution space
to non-negative solutions.}
}
\value{
A tibble containing the debiased Welch estimate
}
\description{
Calculate the Debiased Welch Estimator (slow)
}
