% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_centres.R
\name{get_centres}
\alias{get_centres}
\title{Calculate locations of dwelch centres}
\usage{
get_centres(l, k, delta = 1)
}
\arguments{
\item{l}{length of blocks}

\item{k}{number of bases}

\item{delta}{sampling interval}
}
\value{
A list containing the centres and the bin width
}
\description{
Calculate locations of dwelch centres
}
