% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nfreq.R
\name{get_nfreq}
\alias{get_nfreq}
\title{Calculate number of frequencies exluding zero and Nyquist}
\usage{
get_nfreq(l)
}
\arguments{
\item{l}{length of blocks}
}
\value{
An integer of number of frequencies
}
\description{
Calculate number of frequencies exluding zero and Nyquist
}
