% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whittle.R
\name{whittle}
\alias{whittle}
\title{Whittle Likelihood}
\usage{
whittle(psd, periodogram, return_log = TRUE)
}
\arguments{
\item{psd}{Power spectral density}

\item{periodogram}{Periodogram}

\item{return_log}{Return the log likelihood}
}
\value{
The value of the Whittle likelihood
}
\description{
Whittle Likelihood
}
