% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bochner.R
\name{bochner}
\alias{bochner}
\title{Compute Bochner's Theorem}
\usage{
bochner(acf, delta = 1, h = NULL, full = FALSE)
}
\arguments{
\item{acf}{acf of the time-series}

\item{delta}{sampling interval}

\item{h}{optional window to apply to data. Defaults to rectangular}

\item{full}{boolean value if return full symmetric spectrum}
}
\value{
A vector of the positive spectral density,
excluding zero and Nyquist frequencies
}
\description{
Compute Bochner's Theorem
}
