% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_bases.R
\name{build_bases}
\alias{build_bases}
\title{Calculate biased bases (slow)}
\usage{
build_bases(
  l,
  k = NULL,
  h = NULL,
  delta = 1,
  centres = NULL,
  widths = NULL,
  lowers = NULL,
  uppers = NULL
)
}
\arguments{
\item{l}{length of blocks}

\item{k}{number of bases}

\item{h}{optional window to apply to data. Defaults to rectangular}

\item{delta}{sampling interval}
}
\value{
A l x k matrix of each of the bases
}
\description{
Calculate biased bases (slow)
}
