% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwelch.R
\name{dwelch}
\alias{dwelch}
\title{Calculate the Debiased Welch Estimator (slow)}
\usage{
dwelch(ts, m, l, s, k, delta = 1, h = NULL)
}
\arguments{
\item{ts}{time-series as a vector}

\item{m}{number of blocks}

\item{l}{length of blocks}

\item{s}{shift factor}

\item{k}{number of bases}

\item{delta}{sampling interval}

\item{h}{optional window to apply to data. Defaults to rectangular}
}
\value{
A tibble containing the debiased Welch estimate
}
\description{
Calculate the Debiased Welch Estimator (slow)
}
