% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwelch.R
\name{pwelch}
\alias{pwelch}
\title{Calculate Welch's Estimator}
\usage{
pwelch(ts, m, l, s, delta = 1, h = NULL)
}
\arguments{
\item{ts}{time-series as a vector}

\item{m}{number of blocks}

\item{l}{length of blocks}

\item{s}{shift factor}

\item{delta}{sampling interval}

\item{h}{optional window to apply to data. Defaults to rectangular}
}
\value{
A tibble containing Welch's estimate
}
\description{
Calculate Welch's Estimator
}
